requirements.txt
aniso8601
hearthstone==5.51.0
hearthstone-data==124497.1
hslog==1.16.1
MouseInfo
numpy
opencv-python
Pillow
psutil
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
pywin32
PyYAML
PyQt5
pinyin
keyboard